refactor(agent-hooks): extract Codex session start handler
