--- src/main/core/agent-hooks/codex-session-id.ts
@@ -1,7 +1,0 @@
-export function extractCodexProviderSessionId(body: Record<string, unknown>): string | undefined {
-  const candidates = [body.session_id, body.resource_id, body.resourceId, body.sessionId];
-  for (const candidate of candidates) {
-    if (typeof candidate === 'string' && candidate.trim()) return candidate.trim();
-  }
-  return undefined;
-}
